third try

+++ db/migrate/20150408155033_create_vats.rb
@@ -1,7 +1,7 @@
 class CreateVats < ActiveRecord::Migration
   def change
     create_table :vats do |t|
-      t.decimal :value, default: false, null: false
+      t.float :value, default: false, null: false
 
       t.timestamps null: false
     end
